place ts code back in root file
