+++ src/quasar.ts
@@ -8,6 +8,7 @@ import Quasar, {
   QToolbar,
   QToolbarTitle,
   QBtn,
+  QAlert,
   QBtnGroup,
   QIcon,
   QList,
@@ -57,6 +58,7 @@ Vue.use(Quasar, {
     QPage,
     QToolbar,
     QToolbarTitle,
+    QAlert,
     QBtn,
     QBtnGroup,
     QIcon,
